feat: customHook to editPost
